requirements.txt
# Runtime dependencies for the Tractian GTM pipeline.
# Slim list — everything here is actually imported somewhere under src/.

# HTTP + env
requests==2.32.3
python-dotenv==1.1.1

# Progress bar for the orchestrator
tqdm==4.67.1

# Pydantic schemas for pipeline rows and ICP scores
pydantic==2.11.7

# Excel output (XLSX with 3 styled sheets)
openpyxl==3.1.5

# FastAPI backend for the dashboard
fastapi==0.116.1
uvicorn==0.35.0
